split model_data/helpers and extract_apc_data from model_data_extract
