+++ model_data/helpers.py
@@ -1,0 +1,61 @@
+"""Helper methods/tables"""
+
+import pyspark.sql.functions as F
+from databricks.connect import DatabricksSession
+from pyspark.sql import DataFrame, SparkSession
+
+
+def add_tretspef_column(self: DataFrame) -> DataFrame:
+    """Add tretspef column to DataFrame
+
+    :param self: The data frame to add the tretpsef column to
+    :type df: DataFrame
+    :return: The data frame
+    :rtype: DataFrame
+    """
+
+    specialties = [
+        "100",
+        "101",
+        "110",
+        "120",
+        "130",
+        "140",
+        "150",
+        "160",
+        "170",
+        "300",
+        "301",
+        "320",
+        "330",
+        "340",
+        "400",
+        "410",
+        "430",
+        "520",
+    ]
+
+    tretspef_column = (
+        F.when(F.col("tretspef_raw").isin(specialties), F.col("tretspef_raw"))
+        .when(F.expr("tretspef_raw RLIKE '^1(?!80|9[02])'"), F.lit("Other (Surgical)"))
+        .when(
+            F.expr("tretspef_raw RLIKE '^(1(80|9[02])|[2346]|5(?!60)|83[134])'"),
+            F.lit("Other (Medical)"),
+        )
+        .otherwise(F.lit("Other"))
+    )
+
+    return self.withColumn("tretspef", tretspef_column)
+
+
+def get_spark() -> SparkSession:
+    """Get Spark session to use for model data extract
+
+    :return: get the spark context to use
+    :rtype: SparkSession
+    """
+    spark: SparkSession = DatabricksSession.builder.getOrCreate()
+    spark.catalog.setCurrentCatalog("nhp")
+    spark.catalog.setCurrentDatabase("default")
+    spark.conf.set("spark.sql.sources.partitionOverwriteMode", "dynamic")
+    return spark
