+++ model_data/model_data_extract.py
@@ -9,97 +9,7 @@ import sys
 import pyspark.sql.functions as F
 from databricks.connect import DatabricksSession
 from pyspark.sql import DataFrame, SparkSession, Window
-
-
-def add_tretspef_column(self: DataFrame) -> DataFrame:
-    """Add tretspef column to DataFrame
-
-    :param self: The data frame to add the tretpsef column to
-    :type df: DataFrame
-    :return: The data frame
-    :rtype: DataFrame
-    """
-
-    specialties = [
-        "100",
-        "101",
-        "110",
-        "120",
-        "130",
-        "140",
-        "150",
-        "160",
-        "170",
-        "300",
-        "301",
-        "320",
-        "330",
-        "340",
-        "400",
-        "410",
-        "430",
-        "520",
-    ]
-
-    tretspef_column = (
-        F.when(F.col("tretspef_raw").isin(specialties), F.col("tretspef_raw"))
-        .when(F.expr("tretspef_raw RLIKE '^1(?!80|9[02])'"), F.lit("Other (Surgical)"))
-        .when(
-            F.expr("tretspef_raw RLIKE '^(1(80|9[02])|[2346]|5(?!60)|83[134])'"),
-            F.lit("Other (Medical)"),
-        )
-        .otherwise(F.lit("Other"))
-    )
-
-    return self.withColumn("tretspef", tretspef_column)
-
-
-def extract_apc_data(spark: SparkSession, save_path: str, fyear: int) -> None:
-    """Extract APC (+mitigators) data
-
-    :param spark: the spark context to use
-    :type spark: SparkSession
-    :param save_path: where to save the parquet files
-    :type save_path: str
-    :param fyear: what year to extract
-    :type fyear: int
-    """
-    apc = (
-        spark.read.table("apc")
-        .filter(F.col("fyear") == fyear)
-        .withColumnRenamed("epikey", "rn")
-        .withColumnRenamed("provider", "dataset")
-        .withColumn("tretspef_raw", F.col("tretspef"))
-        .withColumn("fyear", F.floor(F.col("fyear") / 100))
-        .withColumn("sex", F.col("sex").cast("int"))
-        .withColumn("sushrg_trimmed", F.expr("substring(sushrg, 1, 4)"))
-    )
-
-    apc = add_tretspef_column(apc)
-
-    (
-        apc.repartition(1)
-        .write.mode("overwrite")
-        .partitionBy(["fyear", "dataset"])
-        .parquet(f"{save_path}/ip")
-    )
-
-    for k, v in [
-        ("activity_avoidance", "activity_avoidance"),
-        ("efficiencies", "efficiency"),
-    ]:
-        (
-            spark.read.table("apc_mitigators")
-            .filter(F.col("type") == v)
-            .drop("type")
-            .withColumnRenamed("epikey", "rn")
-            .join(apc, "rn", "inner")
-            .select("dataset", "fyear", "rn", "strategy", "sample_rate")
-            .repartition(1)
-            .write.mode("overwrite")
-            .partitionBy(["fyear", "dataset"])
-            .parquet(f"{save_path}/ip_{k}_strategies")
-        )
+from model_data.helpers import add_tretspef_column
 
 
 def extract_opa_data(spark: SparkSession, save_path: str, fyear: int) -> None:
@@ -411,7 +321,6 @@ def main(save_path: str, fyear: int) -> None:
 
     spark.conf.set("spark.sql.sources.partitionOverwriteMode", "dynamic")
 
-    extract_apc_data(spark, save_path, fyear)
     extract_opa_data(spark, save_path, fyear)
     extract_ecds_data(spark, save_path, fyear)
     extract_birth_factors_data(spark, save_path, fyear)
